separated commands and actions from main.go, created delete command
